+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # AWS Mobile SDK for iOS CHANGELOG
 
 ## Unreleased
--Features for next release
+
+### New features
+
+- **Amazon S3**
+  - The AWSS3 SDK now uses [virtual-host style addressing](https://docs.aws.amazon.com/AmazonS3/latest/dev/VirtualHosting.html#virtual-hosted-style-access) for DNS-compliant bucket names. (See [issue #1535](https://github.com/aws-amplify/aws-sdk-ios/issues/1535), and [PR #2996](https://github.com/aws-amplify/aws-sdk-ios/pull/2996)). This does not affect AWSS3TransferUtility, which already uses virtual-host style addressing in its operations.
 
 ## 2.16.0